SAGE-1586: fix chirpstack postgresql and redis deployment
The postgresql and redis deployments should actually be statefulsets
as they deal with persistant data. To correct this a one-time operation
is added to remove the old deployments and associated data.

Additionaly, as part of this fix the postgresql configmap is now
configured as a kustomize configmap.

Also, update start-up probes to be more re-active. This also forces the
server to restart, which we need with a new postgresql database.

Additionally, combined the PVCs into the StatefulSet definitions.